+++ Runtime/Extensions/RaycastHit3DExtension.cs
@@ -55,6 +55,35 @@ namespace ActionCode.ColliderAdapter
         }
 
         /// <summary>
+        /// Draws a 3D CapsuleCast hit using the given params.
+        /// </summary>
+        /// <param name="hit"></param>
+        /// <param name="center">The capsule-cast center.</param>
+        /// <param name="axisDirection">The capsule axis direction.</param>
+        /// <param name="direction">The capsule-cast direction.</param>
+        /// <param name="orientation">The capsule-cast orientation.</param>
+        /// <param name="distance">The capsule-cast max distance.</param>
+        /// <param name="diameter">The diameter of the sphere, measured in the object's local space.</param>
+        public static void DrawCapsuleCast(this RaycastHit hit, Vector3 center,
+            Vector3 axisDistance, Vector3 axisDirection,
+            Vector3 rightDirection, Vector3 direction,
+            Quaternion orientation, float distance, float diameter)
+        {
+            var end = center + direction * distance;
+            var color = ExtensionConstants.COLLISION_OFF;
+
+            if (hit.collider)
+            {
+                color = ExtensionConstants.COLLISION_ON;
+                hit.point.Draw(color, ExtensionConstants.POINT_SIZE);
+            }
+
+            Debug.DrawLine(center, end, color);
+            ShapeDebug.DrawCapsule3D(end + axisDistance, end - axisDistance, axisDirection,
+                rightDirection, orientation, radius: diameter * 0.5F, color);
+        }
+
+        /// <summary>
         /// Draws a SphereCast hit using the given params.
         /// </summary>
         /// <param name="hit"></param>